Test each grid module through its own interface
The ensure_grid_files tests sat in test_grids, so the three source modules had
one test module between them. Now test_gridfiles asserts which assets a match
selects and what the caps refuse, and test_grids asserts what a Dataset comes
back looking like - through open_dataset and to_zarr only.
