+++ tests/test_grids.py
@@ -1,16 +1,11 @@
 """Tests for the gridded read path (foehn.open_dataset)."""
 
-from pathlib import Path
 from unittest.mock import patch
 
 import pytest
 
 import foehn
 from foehn.api import open_dataset, to_zarr
-from foehn.collections import COLLECTIONS
-from foehn.fetch import FetchError
-from foehn.gridfiles import ensure_grid_files
-from foehn.workspace import Workspace
 from tests.fakes import InMemoryFetcher
 
 
@@ -62,303 +57,6 @@ def test_open_radar_without_match_raises():
     pytest.importorskip("pyproj")
     with pytest.raises(ValueError, match="match="):
         open_dataset("radar_precip")
-
-
-def test_ensure_netcdf_files_raises_when_no_nc_assets(fetcher, tmp_path):
-    """A collection that only exposes GeoTIFF/ZIP should raise a clear error."""
-    fetcher.any_items = [
-        {"assets": {"a": {"href": "https://data.geo.admin.ch/x/grid.tif"}}},
-        {"assets": {"b": {"href": "https://data.geo.admin.ch/x/bundle.zip"}}},
-    ]
-    with pytest.raises(ValueError, match=r"No \.nc assets"):
-        ensure_grid_files("climate_normals_grid", Workspace(tmp_path), fetcher=fetcher)
-
-
-def test_ensure_netcdf_files_match_selects_subset_via_remote(tmp_path):
-    """A NetCDF match consults the remote listing (to verify completeness) and keeps the subset."""
-    out_dir = tmp_path / "bronze" / "surface_derived_grid"
-    out_dir.mkdir(parents=True)
-    keep = out_dir / "x.rhiresd_ch01h.nc"
-    drop = out_dir / "x.ranomm9120_ch01r.nc"
-    keep.write_bytes(b"x")
-    drop.write_bytes(b"x")
-
-    items = _items_for("x.rhiresd_ch01h.nc", "x.ranomm9120_ch01r.nc")
-    fake = _fake(items)
-    result = ensure_grid_files("surface_derived_grid", Workspace(tmp_path), match="rhiresd", fetcher=fake)
-    assert len(fake.listings) == 1  # multi-file format always verifies against remote
-    assert fake.streams == []  # both already cached
-    assert result == [keep]
-
-
-def test_ensure_netcdf_files_match_downloads_missing_from_partial_cache(tmp_path):
-    """A multi-file NetCDF match must not return a partial cache — it fetches what's missing."""
-    out_dir = tmp_path / "bronze" / "surface_derived_grid"
-    out_dir.mkdir(parents=True)
-    (out_dir / "rhiresd_part1.nc").write_bytes(b"x")  # interrupted earlier download left only part 1
-
-    items = _items_for("rhiresd_part1.nc", "rhiresd_part2.nc")
-
-    def fake_download(_href, filepath):
-        Path(filepath).write_bytes(b"y")
-
-    fake = _fake(items)
-    fake.stream_hook = fake_download
-    result = ensure_grid_files("surface_derived_grid", Workspace(tmp_path), match="rhiresd", fetcher=fake)
-    assert len(fake.streams) == 1  # the missing part 2 is fetched
-    assert {p.name for p in result} == {"rhiresd_part1.nc", "rhiresd_part2.nc"}
-
-
-def test_ensure_grid_files_fetches_missing_concurrently(tmp_path):
-    """Missing files are fetched in parallel, not one after another."""
-    import threading
-
-    out_dir = tmp_path / "bronze" / "surface_derived_grid"
-    out_dir.mkdir(parents=True)
-    names = [f"rhiresd_part{i}.nc" for i in range(6)]
-    items = _items_for(*names)
-
-    active = 0
-    peak = 0
-    lock = threading.Lock()
-    barrier = threading.Barrier(len(names), timeout=5)
-
-    def fake_download(_href, filepath):
-        nonlocal active, peak
-        with lock:
-            active += 1
-            peak = max(peak, active)
-        # Blocks until every download is in flight — times out if they are serial.
-        barrier.wait()
-        with lock:
-            active -= 1
-        Path(filepath).write_bytes(b"y")
-
-    fake = _fake(items)
-    fake.stream_hook = fake_download
-    result = ensure_grid_files("surface_derived_grid", Workspace(tmp_path), match="rhiresd", fetcher=fake)
-
-    assert len(fake.streams) == len(names)
-    assert peak > 1  # genuinely overlapping
-    assert {p.name for p in result} == set(names)
-
-
-def test_ensure_grid_files_deduplicates_repeated_asset(tmp_path):
-    """One file listed under several asset keys is downloaded once.
-
-    Two workers streaming into the same ``.part`` would corrupt it, so the
-    dedupe matters now that these downloads run concurrently.
-    """
-    items = [
-        {
-            "assets": {
-                "data": {"href": "https://data.geo.admin.ch/x/rhiresd_ch01h.nc"},
-                "alternate": {"href": "https://data.geo.admin.ch/x/rhiresd_ch01h.nc"},
-            }
-        }
-    ]
-
-    def fake_download(_href, filepath):
-        Path(filepath).write_bytes(b"y")
-
-    fake = _fake(items)
-    fake.stream_hook = fake_download
-    result = ensure_grid_files("surface_derived_grid", Workspace(tmp_path), match="rhiresd", fetcher=fake)
-
-    assert len(fake.streams) == 1
-    assert [p.name for p in result] == ["rhiresd_ch01h.nc"]
-
-
-def test_grid_listing_accepts_a_cached_walk(tmp_path):
-    """The read path lists on every open, so it opts into a cached listing.
-
-    Whether an entry is reused, and for how long, is the fetcher's business —
-    the TTL is derived from what the walk cost, which only it can measure (see
-    tests/test_fetch.py). What grids decides is that staleness is acceptable
-    here at all; the download paths never ask for it.
-    """
-    out_dir = tmp_path / "bronze" / "surface_derived_grid"
-    out_dir.mkdir(parents=True)
-    (out_dir / "x.rhiresd_ch01h.nc").write_bytes(b"x")
-    fake = _fake(_items_for("x.rhiresd_ch01h.nc"))
-
-    ensure_grid_files("surface_derived_grid", Workspace(tmp_path), match="rhiresd", fetcher=fake)
-
-    assert [cache for _, cache, _, _ in fake.listings] == [True]
-
-
-def test_run_datetime_filter_reads_the_run_stamp():
-    """A match naming a run becomes a STAC datetime; anything else stays unfiltered."""
-    from foehn.gridfiles import _run_datetime_filter
-
-    assert _run_datetime_filter("202605231500-0-t_2m-ctrl") == "2026-05-23T15:00:00Z"
-    # no parseable run stamp, an impossible one, or no match at all
-    assert _run_datetime_filter("t_2m-ctrl") is None
-    assert _run_datetime_filter("cpc26130") is None
-    assert _run_datetime_filter("209913451500") is None
-    assert _run_datetime_filter(None) is None
-
-
-def test_only_grib2_rows_narrow_by_run():
-    """Whether an item's datetime *is* the run is a row, not a test inside the filter.
-
-    CSV and radar collections set item datetime to a catalog-refresh time, so
-    filtering on a real data date there would match nothing — those rows opt out,
-    which is why a 12-digit stamp in a radar match is never sent as a filter.
-    """
-    from foehn import registry
-
-    assert registry.spec("forecast_icon_ch1").grid.run_datetime is True
-    assert registry.spec("analysis_kenda_ch1").grid.run_datetime is True
-    assert registry.spec("radar_precip").grid.run_datetime is False
-    assert registry.spec("surface_derived_grid").grid.run_datetime is False
-
-
-def test_ensure_grid_files_narrows_forecast_listing_by_run(tmp_path):
-    """A GRIB2 open passes the run through as a server-side datetime filter."""
-    out_dir = tmp_path / "bronze" / "forecast_icon_ch1"
-    out_dir.mkdir(parents=True)
-    name = "icon-ch1-eps-202605231500-0-t_2m-ctrl.grib2"
-    (out_dir / name).write_bytes(b"x")
-
-    fake = _fake(_items_for(name))
-    ensure_grid_files(
-        "forecast_icon_ch1",
-        Workspace(tmp_path),
-        suffixes=(".grib2",),
-        match="202605231500-0-t_2m-ctrl",
-        max_files=1,
-        run_datetime=True,
-        fetcher=fake,
-    )
-
-    assert [dt for _, _, dt, _ in fake.listings] == ["2026-05-23T15:00:00Z"]
-
-
-def test_ensure_grid_files_falls_back_when_filtered_listing_misses(tmp_path):
-    """A filtered miss retries unfiltered rather than reporting the file missing."""
-    out_dir = tmp_path / "bronze" / "forecast_icon_ch1"
-    out_dir.mkdir(parents=True)
-    name = "icon-ch1-eps-202605231500-0-t_2m-ctrl.grib2"
-    (out_dir / name).write_bytes(b"x")
-
-    class _FilteredMiss(InMemoryFetcher):
-        """Server-side filter finds nothing; the full walk does."""
-
-        def items(self, collection_id, *, cache=False, datetime_filter=None, max_items=None):
-            found = super().items(collection_id, cache=cache, datetime_filter=datetime_filter, max_items=max_items)
-            return [] if datetime_filter else found
-
-    fake = _FilteredMiss()
-    fake.any_items = _items_for(name)
-    fake.default_body = b"x"
-    result = ensure_grid_files(
-        "forecast_icon_ch1",
-        Workspace(tmp_path),
-        suffixes=(".grib2",),
-        match="202605231500-0-t_2m-ctrl",
-        max_files=1,
-        run_datetime=True,
-        fetcher=fake,
-    )
-    assert [p.name for p in result] == [name]
-    assert len(fake.listings) == 2  # filtered, then the unfiltered retry
-
-
-def test_ensure_single_file_match_validated_against_remote_not_cache(tmp_path):
-    """A single-file match unique in a sparse cache but ambiguous in the collection is rejected."""
-    out_dir = tmp_path / "bronze" / "forecast_icon_ch1"
-    out_dir.mkdir(parents=True)
-    (out_dir / "icon-ch1-eps-202605231500-0-t_2m-ctrl.grib2").write_bytes(b"x")  # only one cached
-
-    # ...but the collection has TWO files matching "t_2m-ctrl".
-    items = _items_for(
-        "icon-ch1-eps-202605231500-0-t_2m-ctrl.grib2",
-        "icon-ch1-eps-202605231800-0-t_2m-ctrl.grib2",
-    )
-    fake = _fake(items)
-    with pytest.raises(ValueError, match="one file at a time"):
-        ensure_grid_files(
-            "forecast_icon_ch1",
-            Workspace(tmp_path),
-            suffixes=(".grib2", ".grib"),
-            match="t_2m-ctrl",
-            max_files=1,
-            fetcher=fake,
-        )
-    assert fake.streams == []  # rejected before any download
-
-
-def test_ensure_single_file_serves_cache_when_remote_unique(tmp_path):
-    """When the collection has exactly one match, the cached file is reused (no re-download)."""
-    out_dir = tmp_path / "bronze" / "forecast_icon_ch1"
-    out_dir.mkdir(parents=True)
-    f = out_dir / "icon-ch1-eps-202605231500-0-t_2m-ctrl.grib2"
-    f.write_bytes(b"x")
-
-    items = _items_for("icon-ch1-eps-202605231500-0-t_2m-ctrl.grib2")
-    fake = _fake(items)
-    result = ensure_grid_files(
-        "forecast_icon_ch1",
-        Workspace(tmp_path),
-        suffixes=(".grib2", ".grib"),
-        match="t_2m-ctrl",
-        max_files=1,
-        fetcher=fake,
-    )
-    assert fake.streams == []
-    assert result == [f]
-
-
-def test_ensure_netcdf_files_match_no_remote_match_raises(tmp_path):
-    fake = _fake([{"assets": {"a": {"href": "https://data.geo.admin.ch/x/ranomm9120.nc"}}}])
-    with pytest.raises(ValueError, match="matching 'rhiresd'"):
-        ensure_grid_files("surface_derived_grid", Workspace(tmp_path), match="rhiresd", fetcher=fake)
-
-
-def test_ensure_netcdf_files_unfiltered_consults_remote_and_downloads_missing(tmp_path):
-    """An unfiltered call must enumerate remote and fetch files missing from a partial cache."""
-    base = tmp_path / "bronze" / "surface_derived_grid"
-    base.mkdir(parents=True)
-    cached = base / "ogd.rhiresd.nc"
-    cached.write_bytes(b"x")  # simulate a prior filtered (match="rhiresd") download
-
-    items = [
-        {"assets": {"a": {"href": "https://data.geo.admin.ch/x/ogd.rhiresd.nc"}}},
-        {"assets": {"b": {"href": "https://data.geo.admin.ch/x/ogd.tabsd.nc"}}},
-    ]
-
-    def fake_download(_href, filepath):
-        Path(filepath).write_bytes(b"y")
-
-    fake = _fake(items)
-    fake.stream_hook = fake_download
-    result = ensure_grid_files("surface_derived_grid", Workspace(tmp_path), fetcher=fake)
-
-    assert len(fake.listings) == 1
-    assert len(fake.streams) == 1  # only the missing file is fetched; cache reused
-    assert {p.name for p in result} == {"ogd.rhiresd.nc", "ogd.tabsd.nc"}
-
-
-def test_ensure_netcdf_files_offline_falls_back_to_cache(tmp_path):
-    """If the STAC API is unreachable, fall back to the cached files with a warning."""
-    base = tmp_path / "bronze" / "surface_derived_grid"
-    base.mkdir(parents=True)
-    (base / "a.nc").write_bytes(b"x")
-
-    fake = _fake()
-    fake.fail(COLLECTIONS["surface_derived_grid"], FetchError("offline"))
-    with pytest.warns(UserWarning, match="without checking the collection"):
-        result = ensure_grid_files("surface_derived_grid", Workspace(tmp_path), fetcher=fake)
-    assert [p.name for p in result] == ["a.nc"]
-
-
-def test_ensure_netcdf_files_offline_no_cache_reraises(tmp_path):
-    """Offline with an empty cache must surface the network error, not swallow it."""
-    fake = _fake()
-    fake.fail(COLLECTIONS["surface_derived_grid"], FetchError("offline"))
-    with pytest.raises(FetchError):
-        ensure_grid_files("surface_derived_grid", Workspace(tmp_path), fetcher=fake)
 
 
 def test_open_dataset_reads_local_netcdf(fetcher, tmp_path):
